Refactoring, extract classes, javadocs
